FIX : Warnings fixed
svn path=/trunk/; revision=634

+++ Source/ZECore/ZECore.cpp
@@ -617,7 +617,7 @@ void ZECore::MainLoop()
 	{
 		QueryPerformanceCounter(&PerformanceCount);
 
-		FrameTime = (float)((PerformanceCount.QuadPart - OldPerformanceCount.QuadPart) / PerformanceCounterFreq.QuadPart);
+		FrameTime = (float)((double)(PerformanceCount.QuadPart - OldPerformanceCount.QuadPart) / (double)PerformanceCounterFreq.QuadPart);
 		OldPerformanceCount = PerformanceCount;
 	}
